@@ -1,12 +1,12 @@
 {
   "name": "@deuex-solutions/react-tour",
-  "version": "1.2.7",
+  "version": "1.2.8",
   "description": "To learn user about how component works in the react app.",
   "main": "dist/reacttour.min.js",
   "types": "index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --mode development --open --hot",
+    "start": "webpack serve --mode development --open --hot",
     "build": "webpack --mode production --config webpack.prod.config.js",
     "build-dev": "webpack --mode development"
   },
@@ -33,27 +33,26 @@
     "scrollparent": "^2.0.1"
   },
   "peerDependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "styled-components": "^6.1.8"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-transform-class-properties": "^7.23.3",
     "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.5",
-    "babel-polyfill": "^6.26.0",
-    "css-loader": "^3.2.0",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "style-loader": "^1.0.0",
-    "styled-components": "^6.1.8",
-    "web-dev-server": "^1.6.7",
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.2.0",
-    "webpack-dev-server": "^3.9.0"
+    "babel-loader": "^9.1.3",
+    "core-js": "^3.36.0",
+    "css-loader": "^6.10.0",
+    "html-loader": "^5.0.0",
+    "html-webpack-plugin": "^5.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "style-loader": "^3.3.4",
+    "styled-components": "^6.1.18",
+    "webpack": "^5.90.3",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.2"
   }
 }
