@@ -1,46 +1,63 @@
 const path = require('path');
-const HtmlWebPackPlugin = require('html-webpack-plugin')
+const HtmlWebPackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-    entry: ['babel-polyfill', './demo/index.js'],
-    devtool: 'sourcemap',
+    entry: ['core-js/stable', './demo/index.js'],
+    devtool: 'source-map',
     devServer: {
-        host: 'localhost'
+        host: 'localhost',
+        static: {
+            directory: path.join(__dirname, 'dist'),
+        },
+        hot: true,
+        open: true,
     },
     module: {
         rules: [
             {
                 test: /\.js$/,
-                exclude: [/node_modules/],
-                use:{
-                    loader: "babel-loader",
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
                 },
             },
             {
                 test: /\.html$/,
                 use: [
-                  {
-                    loader: 'html-loader',
-                    options: { minimize: true },
-                  },
+                    {
+                        loader: 'html-loader',
+                        options: {
+                            minimize: true,
+                            esModule: false,
+                        },
+                    },
                 ],
             },
             {
                 test: /\.css$/,
                 use: ['style-loader', 'css-loader'],
             },
-        ]
+            {
+                test: /\.(png|jpg|jpeg|gif|ico|svg)$/i,
+                type: 'asset/resource',
+                generator: {
+                    filename: 'assets/[name][ext]',
+                },
+            },
+        ],
     },
     plugins: [
         new HtmlWebPackPlugin({
-          template: './demo/public/index.html',
-          filename: './index.html',
+            template: './demo/public/index.html',
+            filename: './index.html',
         }),
     ],
     output: {
         path: path.resolve(__dirname, 'dist/'),
-        publicPath: '',
+        publicPath: '/',
         filename: 'reacttour.js',
-        libraryTarget: 'umd'
-    }
+        library: {
+            type: 'umd',
+        },
+    },
 };
